split demo_test.go into demo_test.go and main_test.go
